1_API/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
greenlet==2.0.2
idna==3.4
numpy==1.24.2
pandas==2.0.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
